// src/app/models/advanced-search/search-conditions.enum.ts
export enum SearchConditionsEnum {
  Equals = 'Equals',
  NotEqual = 'Not Equal To',
  IsMissing = 'Is Missing',
  Contains = 'Contains',
  NotContains = 'Not Contains',
  StartsWith = 'Starts With',
  EndsWith = 'Ends With',
  GreaterThan = 'Greater Than',
  GreaterThanOrEqual = 'Greater Than or Equal To',
  LessThan = 'Less Than',
  LessThanOrEqual = 'Less Than or Equal To',
  InRange = 'Between',
  InRangeDays = 'Between Days',
  InRangeMonths = 'Between Months',
  GreaterThanDays = 'Greater Than Days',
  GreaterThanMonths = 'Greater Than Months',
  LessThanDays = 'Less Than Days',
  LessThanMonths = 'Less Than Months',
  In = 'In',
  NotIn = 'Not In',
}